limpando diretorio: movendo command para main.c
