--- include/command.h
@@ -1,7 +1,0 @@
-#ifndef COMMAND_H
-#define COMMAND_H
-#include "utils.h"
-
-u8 handleCommand(u8* cmd);
-
-#endif // COMMAND_H
